Added LoginPageSteps / small refactor to other relevant classes
