+++ src/test/java/temporary/tests/TestCartPage.java
@@ -4,16 +4,15 @@ import org.junit.After;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import pages.cartPage.CartPage;
-import steps.MainPageSteps;
+import steps.BaseSteps;
 
 import static engine.drivers.WebDriverFactory.quitWebDriver;
+import static steps.BaseSteps.*;
 import static steps.CartPageSteps.goToCheckout;
-import static steps.MainPageSteps.*;
 
 public class TestCartPage extends BaseTest {
 
     protected static CartPage cartPage;
-    protected static MainPageSteps mainPageSteps;
 
     @BeforeClass
     public static void setUp() {
@@ -36,7 +35,7 @@ public class TestCartPage extends BaseTest {
         goToCheckout();
 
         // Then:
-        MainPageSteps.validateURL("https://automationteststore.com/index.php?rt=account/login");
+        BaseSteps.validateURL("https://automationteststore.com/index.php?rt=account/login");
     }
 
 }
